refactor(cms): use a single page per action slug
instead showing a single place tto add the content of all the actions, this commit creates different
cards (pages) within the actions collection in order to edit each action page by slug. If a new slug
is added we need to manually add it within the CMS as this commit does it.

re #72